+++ runtests.py
@@ -1,24 +1,67 @@
 #!/usr/bin/env python
 import sys
 import django
-from django.conf import settings, global_settings as default_settings
 from django.core.management import execute_from_command_line
+from django.conf import settings, global_settings as default_settings
+from os import path
+
+# Give feedback on used versions
+sys.stderr.write('Using Python version {0} from {1}\n'.format(sys.version[:5], sys.executable))
+sys.stderr.write('Using Django version {0} from {1}\n'.format(
+    django.get_version(),
+    path.dirname(path.abspath(django.__file__)))
+)
 
 if not settings.configured:
+    module_root = path.dirname(path.realpath(__file__))
+
+    if django.VERSION >= (1, 8):
+        template_settings = dict(
+            TEMPLATES = [
+                {
+                    'BACKEND': 'django.template.backends.django.DjangoTemplates',
+                    'DIRS': (),
+                    'OPTIONS': {
+                        'loaders': (
+                            'django.template.loaders.filesystem.Loader',
+                            'django.template.loaders.app_directories.Loader',
+                        ),
+                        'context_processors': (
+                            'django.template.context_processors.debug',
+                            'django.template.context_processors.i18n',
+                            'django.template.context_processors.media',
+                            'django.template.context_processors.request',
+                            'django.template.context_processors.static',
+                            'django.contrib.auth.context_processors.auth',
+                        ),
+                    },
+                },
+            ]
+        )
+    else:
+        template_settings = dict(
+            TEMPLATE_DEBUG = True,
+            TEMPLATE_LOADERS = (
+                'django.template.loaders.app_directories.Loader',
+                'django.template.loaders.filesystem.Loader',
+            ),
+            TEMPLATE_CONTEXT_PROCESSORS = list(default_settings.TEMPLATE_CONTEXT_PROCESSORS) + [
+                'django.core.context_processors.request',
+            ],
+        )
+
     settings.configure(
+        DEBUG = False,  # will be False anyway by DjangoTestRunner.
         DATABASES = {
-            'default': {'ENGINE': 'django.db.backends.sqlite3', 'NAME': ':memory:',},
+            'default': {
+                'ENGINE': 'django.db.backends.sqlite3',
+                'NAME': ':memory:'
+            }
         },
         SITE_ID = 1,
-        TEMPLATE_LOADERS = (
-            'django.template.loaders.app_directories.Loader',
-        ),
-        TEMPLATE_CONTEXT_PROCESSORS = list(default_settings.TEMPLATE_CONTEXT_PROCESSORS) + [
-            'django.core.context_processors.request',
-        ],
         INSTALLED_APPS = (
-            'django.contrib.contenttypes',
             'django.contrib.auth',
+            'django.contrib.contenttypes',
             'django.contrib.sites',
             'django.contrib.admin',
             'django.contrib.messages',
@@ -32,8 +75,10 @@ if not settings.configured:
         TEST_RUNNER = 'django.test.simple.DjangoTestSuiteRunner' if django.VERSION < (1,6) else 'django.test.runner.DiscoverRunner',
         ROOT_URLCONF = 'fluentcms_contactform.tests.urls',
         CRISPY_TEMPLATE_PACK = 'bootstrap3',
-        STATIC_URL = '/static/'
+        STATIC_URL = '/static/',
+        **template_settings
     )
+
 
 DEFAULT_TEST_APPS = [
     'fluentcms_contactform',